+++ src/helpers/get-click-within-element.js
@@ -1,0 +1,9 @@
+const getClickWithinElement = (event) => {
+  const rect = event.target.getBoundingClientRect();
+  const x = event.clientX - rect.left;
+  const y = event.clientY - rect.top;
+
+  return { x, y };
+};
+
+export default getClickWithinElement;