+++ src/helpers/index.js
@@ -1,0 +1,5 @@
+import getClickWithinElement from "./get-click-within-element";
+
+export default {
+  getClickWithinElement
+};