+++ src/index.js
@@ -1,1 +1,1 @@
-import './tooltip';
+import './image-paster';